+++ hyperdbg/hyperkd/code/debugger/core/Debugger.c
@@ -357,11 +357,6 @@ DebuggerUninitialize()
     ProcessorsCount = KeQueryActiveProcessorCount(0);
 
     //
-    // Uninitialize the HyperTrace (if it was initialized)
-    //
-    HyperTraceUnInit();
-
-    //
     // Free the Pool manager
     //
     PoolManagerUninitialize();