+++ hyperdbg/libhyperdbg/code/debugger/commands/debugging-commands/load.cpp
@@ -15,6 +15,7 @@
 // Global Variables
 //
 extern BOOLEAN g_IsConnectedToHyperDbgLocally;
+extern BOOLEAN g_IsKdModuleLoaded;
 extern BOOLEAN g_IsVmmModuleLoaded;
 extern BOOLEAN g_IsHyperTraceModuleLoaded;
 
@@ -26,14 +27,15 @@ extern BOOLEAN g_IsHyperTraceModuleLoaded;
 VOID
 CommandLoadHelp()
 {
-    ShowMessages("load : installs the drivers and load the modules.\n\n");
+    ShowMessages("load : installs drivers and load modules.\n\n");
 
-    ShowMessages("syntax : \tload [ModuleName (string)]\n");
+    ShowMessages("syntax : \tload [ModuleNameOrAll (string)]\n");
 
     ShowMessages("\n");
     ShowMessages("\t\te.g : load vmm\n");
     ShowMessages("\t\te.g : load kd\n");
     ShowMessages("\t\te.g : load trace\n");
+    ShowMessages("\t\te.g : load all\n");
 }
 
 /**
@@ -65,8 +67,30 @@ CommandLoad(vector<CommandToken> CommandTokens, string Command)
     //
     // Check for the module
     //
-    if (CompareLowerCaseStrings(CommandTokens.at(1), "vmm") ||
-        CompareLowerCaseStrings(CommandTokens.at(1), "vm"))
+    if (CompareLowerCaseStrings(CommandTokens.at(1), "all") || CompareLowerCaseStrings(CommandTokens.at(1), "."))
+    {
+        //
+        // Load aa Modules
+        //
+        ShowMessages("loading the all modules\n");
+
+        if (HyperDbgInstallKdDriver() == 1 || HyperDbgLoadAllModules() == 1)
+        {
+            ShowMessages("failed to install or load drivers\n");
+            return;
+        }
+
+        //
+        // If in vmi-mode then initialize and load symbols (pdb)
+        // for previously downloaded symbols
+        // When the VMM module is loaded, we use the current
+        // process (HyperDbg's process) as the base for user-mode
+        // symbols
+        //
+        SymbolLocalReload(GetCurrentProcessId());
+    }
+    else if (CompareLowerCaseStrings(CommandTokens.at(1), "vmm") ||
+             CompareLowerCaseStrings(CommandTokens.at(1), "vm"))
     {
         //
         // Check to make sure that the driver is not already loaded
@@ -79,7 +103,7 @@ CommandLoad(vector<CommandToken> CommandTokens, string Command)
         }
 
         //
-        // Load VMM Module
+        // Load the VMM Module
         //
         ShowMessages("loading the vmm module\n");
 
@@ -112,11 +136,39 @@ CommandLoad(vector<CommandToken> CommandTokens, string Command)
         }
 
         //
-        // Load HyperTrace Module
+        // Load the HyperTrace Module
         //
         ShowMessages("loading the trace module\n");
 
         if (HyperDbgInstallKdDriver() == 1 || HyperDbgLoadHyperTraceModule() == 1)
+        {
+            ShowMessages("failed to install or load the driver\n");
+            return;
+        }
+    }
+    else if (CompareLowerCaseStrings(CommandTokens.at(1), "kd") ||
+             CompareLowerCaseStrings(CommandTokens.at(1), "dbg") ||
+             CompareLowerCaseStrings(CommandTokens.at(1), "debugger") ||
+             CompareLowerCaseStrings(CommandTokens.at(1), "debug") ||
+             CompareLowerCaseStrings(CommandTokens.at(1), "kerneldebugger") ||
+             CompareLowerCaseStrings(CommandTokens.at(1), "kerneldebug"))
+    {
+        //
+        // Check to make sure that the driver is not already loaded
+        //
+        if (g_IsKdModuleLoaded)
+        {
+            ShowMessages("the kd module is already running, if you use 'load' before, please "
+                         "first unload it using the 'unload' command\n");
+            return;
+        }
+
+        //
+        // Load the KD Module
+        //
+        ShowMessages("loading the kd module\n");
+
+        if (HyperDbgInstallKdDriver() == 1 || HyperDbgLoadKdModule() == 1)
         {
             ShowMessages("failed to install or load the driver\n");
             return;