+++ hyperdbg/libhyperdbg/code/debugger/core/debugger.cpp
@@ -570,13 +570,13 @@ ShowErrorMessage(UINT32 Error)
         break;
 
     case DEBUGGER_ERROR_HYPERTRACE_NOT_INITIALIZED:
-        ShowMessages("err, the HyperTrace module is not loaded and initialized, "
-                     "use the 'load trace' command to load the HyperTrace module  (%x)\n",
+        ShowMessages("err, the hypertrace module is not loaded and initialized, "
+                     "use the 'load trace' command to load the hypertrace module  (%x)\n",
                      Error);
         break;
 
     case DEBUGGER_ERROR_INVALID_HYPERTRACE_OPERATION_TYPE:
-        ShowMessages("err, invalid HyperTrace operation type is specified (%x)\n",
+        ShowMessages("err, invalid hypertrace operation type is specified (%x)\n",
                      Error);
         break;
 
@@ -617,9 +617,9 @@ ShowErrorMessage(UINT32 Error)
         break;
 
     case DEBUGGER_ERROR_VMM_CANNOT_BE_INITIALIZED_IF_HYPERTRACE_IS_LOADED:
-        ShowMessages("err, HyperTrace is already loaded, please unload HyperTrace module using the "
-                     "'unload' command and then load the 'VMM' module. Then you can load HyperTrace "
-                     "after loading the VMM as it is because HyperTrace behaves differently to sync "
+        ShowMessages("err, hypertrace is already loaded, please unload hypertrace module using the "
+                     "'unload' command and then load the 'VMM' module. Then you can load hypertrace "
+                     "after loading the VMM as it is because hypertrace behaves differently to sync "
                      "with VMM modules; it needs to have this notion that it is running within the "
                      "hypervisor, so that is why it needs to be initialized again if the VMM module "
                      "is loaded (%x)\n",